+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [1.29.5] - 2026-02-28
 
+- Filter: combine inverseTags and inverseFlag (#192)
 - Filter: preserve inverse flag filtering with stripFlags (#193)
 - Types: Update Typescript typings (#194)